pyproject.toml
[tool.poetry]
name = "wisedata"
version = "1.0.13"
description = "AI Assistant for Python Data Analytics"
authors = ["WiseData Team <wisedata.help@gmail.com>"]
homepage = "https://www.wisedata.app/"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.0"
python-dotenv = "^1.0.0"
requests = "2.31.0"
retry = "^0.9.2"
urllib3 = "^1.26.6"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
numpy = "^1.25.0"
matplotlib = "3.7.1"
pandas = "^2.0.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
pytest-sugar = "^0.9.4"
pytest-watch = "^4.2.0"
seaborn = "0.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"